*Somewhat major note*
These changes are (primarily) responsive to the observation
that in some *rare* situations we see *significant* divergence
between the mappings achieved by Splash's splice graph
construction and the hits reported by re-running the full
model on the extracted exons.

I think this suggests a HMMER bug (or hive), but could be
a deep issue with something I'm doing.  We're logging
instances of these "discordant" searches in a specific
output file so I can look into them (and determine how
much of a systemic problem they pose).